// packages/Common/UI/Components/Charts/Types/XAxis/XAxisPrecision.ts
enum XAxisPrecision {
  EVERY_SECOND = "second",
  EVERY_FIVE_SECONDS = "fiveSeconds",
  EVERY_TEN_SECONDS = "tenSeconds",
  EVERY_THIRTY_SECONDS = "thirtySeconds",
  EVERY_MINUTE = "minute",
  EVERY_FIVE_MINUTES = "fiveMinutes",
  EVERY_TEN_MINUTES = "tenMinutes",
  EVERY_FIFTEEN_MINUTES = "fifteenMinutes",
  EVERY_THIRTY_MINUTES = "thirtyMinutes",
  EVERY_HOUR = "hour",
  EVERY_TWO_HOURS = "twoHours",
  EVERY_THREE_HOURS = "threeHours",
  EVERY_SIX_HOURS = "sixHours",
  EVERY_TWELVE_HOURS = "twelveHours",
  EVERY_DAY = "day",
  EVERY_TWO_DAYS = "twoDays",
  EVERY_WEEK = "week",
  EVERY_TWO_WEEKS = "twoWeeks",
  EVERY_MONTH = "month",
  EVERY_TWO_MONTHS = "twoMonths",
  EVERY_THREE_MONTHS = "threeMonths",
  EVERY_SIX_MONTHS = "sixMonths",
  EVERY_YEAR = "year",
}

export default XAxisPrecision;